third_party/WebKit/LayoutTests/http/tests/devtools/console-show-all-messages-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that console shows messages only from specific context when show target checkbox is checked.


Running: testInitialState
message from page!
message from iframe!

Running: testPageOnlyMessages
message from page!

Running: testFrameOnlyMessages
message from iframe!

Running: testAllMessagesWithFrameContext
message from page!
message from iframe!
